requirements.txt
streamlit
google-generativeai
gtts
sounddevice
soundfile
numpy
openai-whisper